% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_wilder_kaiser}
\alias{bb_wilder_kaiser}
\title{Bounding box of Wilder Kaiser}
\format{
A numeric vector with 4 elements:
\describe{
\item{xmin}{Minimum longitude}
\item{ymin}{Minimum latitude}
\item{xmax}{Maximum longitude}
\item{ymax}{Maximum latitude}
}
}
\source{
OpenStreetMap data
}
\usage{
bb_wilder_kaiser
}
\description{
A dataset containing the bounding box coordinates of the Wilder Kaiser mountain range.
}
\keyword{datasets}
