% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ridgelines.R
\name{calculate_ridgelines}
\alias{calculate_ridgelines}
\title{Calculate ridgeline data from elevation}
\usage{
calculate_ridgelines(elevation = NULL, n_lines = 30)
}
\arguments{
\item{elevation}{A raster object containing elevation data, or NULL to use last elevation}

\item{n_lines}{Number of ridgelines to calculate}
}
\value{
A data frame containing the ridgeline data with columns x, y, elevation, and group
}
\description{
Takes elevation data and calculates the data needed for plotting ridgelines
by sampling the elevation at regular y-intervals. You will only need this
function if you want to create your own custom plots from the ground up.
}
\seealso{
\code{\link{plot_ridgelines}}
}
