% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ele_wilder_kaiser}
\alias{ele_wilder_kaiser}
\title{Elevation data of Wilder Kaiser}
\format{
A raster object containing elevation data
}
\source{
SRTM elevation data
}
\usage{
ele_wilder_kaiser
}
\description{
A dataset containing elevation data for the Wilder Kaiser mountain range.
}
\keyword{datasets}
