% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get elevation data for a bounding box}
\usage{
get_elevation(bb = NULL, z = 9)
}
\arguments{
\item{bb}{A bounding box object (sf::bbox) or NULL to use the last drawn bounding box}

\item{z}{Zoom level for elevation data (1-14, higher means more detail). Default is 9,
which provides a good balance between detail and download size.}
}
\value{
A raster object containing elevation data
}
\description{
Retrieves elevation data for a specified geographic area using the elevatr package.
The data is returned as a raster object that can be used with the plotting
functions in this package.
}
\examples{
# Using the included Wilder Kaiser bounding box
elevation <- get_elevation(bb_wilder_kaiser)
plot_elevation()

# Using a higher zoom level for more detail
detailed_elevation <- get_elevation(bb_wilder_kaiser, z = 12)
plot_elevation()

}
