% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-last.R
\name{get_last_bb}
\alias{get_last_bb}
\title{Get the last bounding box data}
\usage{
get_last_bb()
}
\description{
Get the last bounding box data
}
\seealso{
\code{\link[=draw_bb]{draw_bb()}}
}
\keyword{internal}
