% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-last.R
\name{get_last_ele}
\alias{get_last_ele}
\title{Get the last elevation data}
\usage{
get_last_ele()
}
\description{
Get the last elevation data
}
\seealso{
\code{\link[=get_elevation]{get_elevation()}}
}
\keyword{internal}
