% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elevation.R
\name{plot_elevation}
\alias{plot_elevation}
\title{Plot elevation data as a heatmap}
\usage{
plot_elevation(elevation = NULL, low_color = "darkblue", high_color = "white")
}
\arguments{
\item{elevation}{A raster object containing elevation data, as returned by get_elevation()}

\item{low_color}{Color for lowest elevations (default: "darkblue")}

\item{high_color}{Color for highest elevations (default: "white")}
}
\value{
A ggplot object that can be further customized using ggplot2 functions
}
\description{
Creates a heatmap visualization of elevation data. This function provides a
quick way to visualize the terrain in your area of interest.
}
\examples{
# Create a basic elevation plot
plot_elevation(ele_wilder_kaiser)

# Customize colors
plot_elevation(ele_wilder_kaiser, low_color = "darkgreen", high_color = "yellow")

}
