% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_solver.R, R/gaussian_solver.R,
%   R/hurdle_zero_inflated.R, R/poisson_solver.R
\name{predict.logistic_ridge}
\alias{predict.logistic_ridge}
\alias{predict.gaussian_ridge}
\alias{predict.inflation_model}
\alias{predict.poisson_ridge}
\title{#' @importFrom stats predict}
\usage{
\method{predict}{logistic_ridge}(object, new_data, ...)

\method{predict}{gaussian_ridge}(object, new_data, ...)

\method{predict}{inflation_model}(object, new_data, ...)

\method{predict}{poisson_ridge}(object, new_data, ...)
}
\arguments{
\item{object}{The fitted model.}

\item{new_data}{New data (always a matrix).}

\item{...}{Additional parameters (only for generic consistency).}
}
\value{
Predictions.
}
\description{
#' @importFrom stats predict

#' @importFrom stats predict

#' @importFrom stats predict
}
