% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Solve ridge regression problems}
\usage{
ridge(
  y,
  X,
  lambda = 1e-08,
  family = c("gaussian", "poisson", "hurdle", "zero-inflated"),
  ...
)
}
\arguments{
\item{y}{The target value - a vector of type numeric.}

\item{X}{The design matrix - this will not be modified in any way.}

\item{lambda}{The ridge penalty parameter - set to NULL to try to estimate.}

\item{family}{Technically allows non-families (such as hurdle/zero-inflated options).
Specifies the distribution of your target.}

\item{...}{Additional parameters passed to solvers (currently not implemented).}
}
\description{
Solve ridge regression problems
}
\details{
No choice of link functions or offsets is provided here - all links are canonical,
all offsets can be provided within **X** (though whether this makes sense is left to the user to decide -
we recommend **not** using this library if you want to use an offset.)
}
