% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-last.R
\name{set_last_bb}
\alias{set_last_bb}
\title{Set the last bounding box data}
\usage{
set_last_bb(value)
}
\description{
Set the last bounding box data
}
\seealso{
\code{\link[=get_last_bb]{get_last_bb()}}
}
\keyword{internal}
