% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-last.R
\name{set_last_ele}
\alias{set_last_ele}
\title{Set the last elevation data}
\usage{
set_last_ele(value)
}
\description{
Set the last elevation data
}
\seealso{
\code{\link[=get_last_ele]{get_last_ele()}}
}
\keyword{internal}
